// src/app/models/plato/plato.interface.ts
export interface Plato {
  id: number;
  Nombre: string;
  Tipo: string;
  Altranuces: boolean;
  Apio: boolean;
  Cacahuetes: boolean;
  Crustaceos: boolean;
  Sulfitos: boolean;
  Cascara: boolean;
  Gluten: boolean;
  Sesamo: boolean;
  Huevo: boolean;
  Lacteos: boolean;
  Moluscos: boolean;
  Mostaza: boolean;
  Pescado: boolean;
  Soja: boolean;
}